for #2457, rewrite.token.impl => rewrite.token.generator
